% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dire_bonjour}
\alias{dire_bonjour}
\title{Fonctions de salutations}
\usage{
dire_bonjour(prenom = "toi")
}
\arguments{
\item{prenom}{Prenom de la personne a saluer}
}
\value{
None
}
\description{
Saluer l'utilisateur par son prenom
}
